% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_a_name.R
\name{draw_a_name}
\alias{draw_a_name}
\title{Given a name, the trend of the number of occurence}
\usage{
draw_a_name(a_name)
}
\arguments{
\item{a_name}{name}
}
\value{
a graph
}
\description{
Given a name, the trend of the number of occurence
}
\examples{
\dontrun{
draw_a_name("the name you want")
}

}
