% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_names.R
\name{draw_names}
\alias{draw_names}
\title{Plot the evolution of names occurences for different names}
\usage{
draw_names(names)
}
\arguments{
\item{names}{names you want to observe occurences data}
}
\value{
graph.
}
\description{
Plot the evolution of names occurences for different names
}
\examples{
\dontrun{

draw_names(c("name1","name2",...))

}
}
