% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Help.R
\docType{package}
\name{myfirstpackage}
\alias{myfirstpackage}
\alias{myfirstpackage-package}
\title{Training on R package}
\description{
functions : plot graphs linked with prenoms dataset  and work with csv and excel files
}
\author{
Amir Benmahjoub <amir.benmahjoub@polytechnique.edu>
}
