% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_as_csv.R
\name{save_as_csv}
\alias{save_as_csv}
\title{Save a data.frame as a csv file}
\usage{
save_as_csv(dataset, filename, row.names = FALSE, ...)
}
\arguments{
\item{dataset}{data.frame}

\item{filename}{a path with ".csv" at the end}
}
\value{
nothing
}
\description{
Save a data.frame as a csv file
}
\examples{
\dontrun{
save_as_csv("name.xls")
}


}
